+++ lib/Dashi/Components/TextField.hs
@@ -28,7 +28,7 @@ instance Token Background where
     tokenAttr (Background state) = tokenAttr state
 
 instance ValueToken Background where
-    type ValueType Background = LightDark (Color (Alpha OKLCH) Float)
+    type ValueType Background = LightDark (Color (Alpha OKLCH) Double)
     tokenValue (Background state) =
         tokenValue (Colour.Text Default) <&> \(ColorOKLCHA l c h _) ->
             ColorOKLCHA l c h $ case state of
